Move BindingsGeneration to a separate file
